Move fractal into lib
Ideally the theme code would be a dependency to avoid it being
confused for GOV.UK Frontend code, but not an issue for the
alpha
