+++ config/paths.json
@@ -16,10 +16,10 @@
   "bundleJs": "dist/bundle/assets/js/",
   "bundleTemplates": "dist/bundle/templates/",
   "bundleComponents": "dist/bundle/components/",
-  "fractal": "fractal",
-  "fractalThemeCss": "fractal/govuk-theme/assets/css/",
-  "fractalThemeScss": "fractal/govuk-theme/assets/scss/",
-  "fractalThemeViews": "fractal/govuk-theme/views/",
+  "fractal": "lib/fractal",
+  "fractalThemeCss": "lib/fractal/govuk-theme/assets/css/",
+  "fractalThemeScss": "lib/fractal/govuk-theme/assets/scss/",
+  "fractalThemeViews": "lib/fractal/govuk-theme/views/",
   "gem": "dist/gem/",
   "gemLib": "dist/gem/lib",
   "gemConfig": "dist/gem/config",
