+++ lib/fractal/config/fractal.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const paths = require('../../config/paths.json')
-const packageJson = require('../../package.json')
+const paths = require('../../../config/paths.json')
+const packageJson = require('../../../package.json')
 // Create a new Fractal instance
 const fractal = require('@frctl/fractal').create()
 const govukFractalTheme = require('../govuk-theme/index.js')
