+++ lib/fractal/extensions/index.js
