+++ lib/fractal/govuk-theme/assets/images/anchored-heading-icon-2x.png
