+++ lib/fractal/govuk-theme/assets/images/anchored-heading-icon.png
