+++ lib/fractal/govuk-theme/assets/images/gov.uk_logotype_crown_invert_trans.png
