+++ lib/fractal/govuk-theme/assets/images/govuk-icn-close.png
