+++ lib/fractal/govuk-theme/assets/images/govuk-icn-close@2x.png
