+++ lib/fractal/govuk-theme/assets/images/govuk-icn-numbered-list.png
