+++ lib/fractal/govuk-theme/assets/scss/_theme-overrides.scss
