+++ lib/fractal/govuk-theme/assets/scss/tech-docs-template/_syntax-highlighting.scss
