+++ lib/fractal/govuk-theme/assets/scss/tech-docs-template/modules/_govuk-logo.scss
