+++ lib/fractal/govuk-theme/assets/scss/tech-docs-template/palette/syntax-highlighting.scss
