+++ lib/fractal/govuk-theme/index.js
@@ -1,4 +1,5 @@
 'use strict'
+const path = require('path')
 
 /* Fractal theme
 ----------------------------------------------------------------------------- */
@@ -36,10 +37,10 @@ const customTheme = mandelbrot({
 })
 
 // Specify a template directory to override any view templates
-customTheme.addLoadPath('./fractal/govuk-theme/views/')
+customTheme.addLoadPath(path.join(__dirname, 'views'))
 
 // Specify the static assets directory that contains the custom stylesheet
-customTheme.addStatic('./fractal/govuk-theme/assets/', '/theme')
+customTheme.addStatic(path.join(__dirname, 'assets'), '/theme')
 
 // Export the customised theme instance so it can be used in Fractal projects
 module.exports = customTheme
