+++ lib/tasks/fractal-assets.js
@@ -21,7 +21,7 @@ gulp.task('fractal:assets:styles', () => {
   return gulp.src(paths.fractalThemeScss + '**/**/*.scss')
    .pipe(sass({
      outputStyle: 'expanded',
-     includePaths: ['./fractal/govuk-theme/assets/scss/tech-docs-template']
+     includePaths: [paths.fractalThemeScss + 'tech-docs-template']
    }).on('error', sass.logError))
     .pipe(gulp.dest(paths.fractalThemeCss))
 })
