+++ lib/tasks/fractal-build.js
@@ -2,7 +2,7 @@
 
 const gulp = require('gulp')
 // require the Fractal instance
-const fractal = require('../../fractal/config/fractal.js')
+const fractal = require('../../lib/fractal/config/fractal.js')
 // keep a reference to the fractal CLI console utility
 const logger = fractal.cli.console
