+++ lib/tasks/fractal.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const gulp = require('gulp')
-const fractal = require('../../fractal/config/fractal.js') // require the Fractal instance
+const fractal = require('../../lib/fractal/config/fractal.js') // require the Fractal instance
 
 function serveDynamicFractal () {
   const port = process.env.PORT ? parseInt(process.env.PORT, 10) : 3000
